CU Boulder BombLab
The point of this assignment is to defuse malicious malware and get flags to prove the success of each stage. If the malware is not defused and you trip bomb to detonate the server is pinged and a mark is scored against you. This involves many break controls and the use of assembly code skills